+++ packages/ionic-react/CHANGELOG.md
@@ -5,6 +5,7 @@
 # Features
 
 - update Ionic to 5.4.1
+- add `ionic.config.json` to application
 
 # BREAKING CHANGES